requirements.txt
PyQt6
yara-python
python-magic
requests
